Move GraphButton to its own file
